+++ library/src/components/filter/default-types/Custom.vue
@@ -1,0 +1,72 @@
+<template lang="pug">
+  .nio-filter-properties.boolean
+    NioFilterProperty(
+      v-bind:value.sync="filter.value"
+      :description="description"
+      :options="filter.options ? filter.options : defaultOptions"
+      :custom-option-loading="false"
+      :filter="filter"
+    )
+      template(
+        v-for="(index, name) in $scopedSlots" 
+        v-slot:[name]="data"
+      )
+        slot(
+          v-bind="data"
+          :name="name" 
+        )   
+</template>
+
+<script>
+
+import NioFilterProperty from '../FilterProperty'
+
+export default {
+  name: 'nio-filter-properties-object',
+  components: { NioFilterProperty },
+  props: {
+    "filter": { type: Object, required: true },
+    "customOptionLoading": { type: Boolean, required: false, default: false }
+  },
+  data: () => ({
+    description: 'Select the data to include'
+  }),	
+  computed: {
+    defaultOptions() {
+      if (!this.filter) {
+        return []
+      }
+      return [
+        {
+          label: 'Optional Attribute',
+          value: 'default',
+        },
+        {
+          label: "Required Attribute",
+          value: 'ifPresent',
+        }
+      ]
+    }
+  },
+  watch: {
+    filter: {
+      deep: true,
+      handler() {
+        this.updateValue()
+      }
+    }
+  },
+  mounted() {
+    this.updateValue()
+  },
+  methods: {
+    updateValue() {
+      const options = this.filter.options ? this.filter.options : this.defaultOptions
+      this.$emit('valueChanged', [options.find(option => option.value === this.filter.value).label])
+    }
+  }
+}
+</script>
+
+<style lang="sass" scoped>
+</style>