+++ library/src/test-components/filters/custom/instances/default.js
@@ -1,0 +1,22 @@
+const Default = { 
+	name: "custom",
+	type: "custom",
+	title: "Default",
+	description: "Default description",
+	value: 'default',
+	options: [
+			{
+			label: `Inlcude all values`,
+			value: 'default',
+		},
+		{
+			label: "Custom",
+			value: 'custom',
+		}
+	],
+	customOption: {
+		value: 'test'
+	}
+}
+
+export default Default