+++ library/src/modules/app/schema/attributeModule.js
@@ -300,5 +300,6 @@ export {
   makeSelected,
   makeSelectedFromSchemaPreset,
 	isExportable,
-	getJoinOptionsByPath
+	getJoinOptionsByPath,
+	makeDotDelimitedPropertyPath
 }