+++ library/src/test-components/TestSelect.vue
@@ -8,6 +8,7 @@
       :label="'Label'"
       key="0"
       fluid-width
+      additional-content-class="test"
     )
     NioSelect.select(
       v-model="selectedItems" 
@@ -117,9 +118,9 @@ export default {
     ],
     selectedItems: ['apple'],
     selectedRegions: [ {
-			name: 'US',
-			value: 1
-		}],
+      name: 'US',
+      value: 1
+    }],
     radio: 'Jacob',
     size: null,
     regions: [
@@ -140,17 +141,17 @@ export default {
         value: 4
       }
     ]  
-	}),
-	mounted() {
-		setTimeout(() => {
-			this.selectedRegions = [
-				{
-					name: 'G',
-					value: 4
-				}
-			]
-		}, 1000);
-	}
+  }),
+  mounted() {
+    setTimeout(() => {
+      this.selectedRegions = [
+        {
+          name: 'G',
+          value: 4
+        }
+      ]
+    }, 1000);
+  }
 };
 </script>