+++ CHANGELOG.md
@@ -130,3 +130,4 @@ Restructured codebase to conform to the Vue Style Guide
 v0.4.6
 Added 'pending' status to condition to display certain subscriptio properties in **NioBuyerStudioSubscription**
 Allow primitive attributes with **is_join_key** to return joinable datasets in attributeModule.getJoinOptionsByPath 
+Added **minCharsToSearch** optional prop to activate **NioSlatTableSearch** on specified number of characters