+++ CHANGELOG.md
@@ -134,4 +134,5 @@ Added **minCharsToSearch** optional prop to activate **NioSlatTableSearch** on s
 Added **custom** filter type with custom filter property slot defeined by **filter-properties-custom-${filterName}**
 Added default slot to **NioFilterGroup**
 Export *makeDotDelimitedPropertyPath** function in **attributeModule.js**
-added **areSamePaths** function to **attributeModule** exports
+Added **areSamePaths** function to **attributeModule** exports
+Added **additionalContentClass** optional prop to **NioTextField**