+++ CHANGELOG.md
@@ -136,3 +136,4 @@ Added default slot to **NioFilterGroup**
 Export *makeDotDelimitedPropertyPath** function in **attributeModule.js**
 Added **areSamePaths** function to **attributeModule** exports
 Added **additionalContentClass** optional prop to **NioTextField**
+Added **NioCombobox** component